Improve the internal processing of messages
