--- app-modules/engagement/src/Filament/Actions/RelationManagerSendEngagementAction.php
@@ -1,430 +1,0 @@
-<?php
-
-/*
-<COPYRIGHT>
-
-    Copyright © 2016-2026, Canyon GBS Inc. All rights reserved.
-
-    Advising App® is licensed under the Elastic License 2.0. For more details,
-    see https://github.com/canyongbs/advisingapp/blob/main/LICENSE.
-
-    Notice:
-
-    - You may not provide the software to third parties as a hosted or managed
-      service, where the service provides users with access to any substantial set of
-      the features or functionality of the software.
-    - You may not move, change, disable, or circumvent the license key functionality
-      in the software, and you may not remove or obscure any functionality in the
-      software that is protected by the license key.
-    - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor’s trademarks is subject
-      to applicable law.
-    - Canyon GBS Inc. respects the intellectual property rights of others and expects the
-      same in return. Canyon GBS® and Advising App® are registered trademarks of
-      Canyon GBS Inc., and we are committed to enforcing and protecting our trademarks
-      vigorously.
-    - The software solution, including services, infrastructure, and code, is offered as a
-      Software as a Service (SaaS) by Canyon GBS Inc.
-    - Use of this software implies agreement to the license terms and conditions as stated
-      in the Elastic License 2.0.
-
-    For more information or inquiries please visit our website at
-    https://www.canyongbs.com or contact us via email at legal@canyongbs.com.
-
-</COPYRIGHT>
-*/
-
-namespace AdvisingApp\Engagement\Filament\Actions;
-
-use AdvisingApp\Engagement\Actions\CreateEngagement;
-use AdvisingApp\Engagement\DataTransferObjects\EngagementCreationData;
-use AdvisingApp\Engagement\Filament\Forms\Components\EngagementSmsBodyInput;
-use AdvisingApp\Engagement\Models\EmailTemplate;
-use AdvisingApp\Engagement\Models\Engagement;
-use AdvisingApp\Notification\Enums\NotificationChannel;
-use AdvisingApp\Prospect\Models\Prospect;
-use AdvisingApp\Prospect\Models\ProspectEmailAddress;
-use AdvisingApp\Prospect\Models\ProspectPhoneNumber;
-use AdvisingApp\StudentDataModel\Enums\EmailAddressOptInOptOutStatus;
-use AdvisingApp\StudentDataModel\Models\Contracts\Educatable;
-use AdvisingApp\StudentDataModel\Models\Student;
-use AdvisingApp\StudentDataModel\Models\StudentEmailAddress;
-use AdvisingApp\StudentDataModel\Models\StudentPhoneNumber;
-use Exception;
-use Filament\Actions\Action;
-use Filament\Actions\CreateAction;
-use Filament\Forms\Components\Checkbox;
-use Filament\Forms\Components\DateTimePicker;
-use Filament\Forms\Components\RichEditor;
-use Filament\Forms\Components\RichEditor\ToolbarButtonGroup;
-use Filament\Forms\Components\Select;
-use Filament\Forms\Components\Toggle;
-use Filament\Forms\Components\ToggleButtons;
-use Filament\Resources\RelationManagers\RelationManager;
-use Filament\Schemas\Components\Actions;
-use Filament\Schemas\Components\Fieldset;
-use Filament\Schemas\Components\Utilities\Get;
-use Filament\Schemas\Components\Utilities\Set;
-use Filament\Schemas\Components\Wizard\Step;
-use Filament\Schemas\Schema;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Query\Expression;
-use Illuminate\Support\Carbon;
-use Spatie\MediaLibrary\MediaCollections\Models\Media;
-
-class RelationManagerSendEngagementAction extends CreateAction
-{
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->icon('heroicon-m-chat-bubble-bottom-center-text')
-            ->label('New')
-            ->slideOver()
-            ->modalHeading('Send Message')
-            ->modalDescription(function (RelationManager $livewire): string {
-                $educatable = $livewire->getOwnerRecord();
-
-                assert($educatable instanceof Educatable);
-
-                $educatableName = $educatable->getAttributeValue($educatable::displayNameKey());
-
-                return "Send an engagement to {$educatableName}.";
-            })
-            ->model(Engagement::class)
-            ->authorize(function (RelationManager $livewire) {
-                $ownerRecord = $livewire->getOwnerRecord();
-
-                return auth()->user()->can('create', [Engagement::class, $ownerRecord instanceof Prospect ? $ownerRecord : null]);
-            })
-            ->disabled(function (RelationManager $livewire): bool {
-                $educatable = $livewire->getOwnerRecord();
-
-                assert($educatable instanceof Educatable);
-
-                return ! $educatable->hasAnyValidContactRoute();
-            })
-            ->tooltip(function (RelationManager $livewire): ?string {
-                $educatable = $livewire->getOwnerRecord();
-
-                assert($educatable instanceof Educatable);
-
-                if (! $educatable->hasAnyValidContactRoute()) {
-                    $label = $educatable::getLabel();
-
-                    return "This {$label} does not have valid contact information in their record.";
-                }
-
-                return null;
-            })
-            ->mountUsing(function (RelationManager $livewire, Schema $schema): void {
-                $educatable = $livewire->getOwnerRecord();
-
-                assert($educatable instanceof Educatable);
-
-                $defaultChannel = $educatable->getDefaultEngagementChannel();
-
-                $schema->fill([
-                    'channel' => $defaultChannel?->value,
-                    'recipient_route_id' => $defaultChannel
-                        ? $educatable->getDefaultRouteForEngagementChannel($defaultChannel)
-                        : null,
-                ]);
-            })
-            ->steps(fn (): array => [
-                Step::make('Recipient Details')
-                    ->schema([
-                        Fieldset::make('Message Type')
-                            ->schema([
-                                ToggleButtons::make('channel')
-                                    ->inline()
-                                    ->options(
-                                        fn (RelationManager $livewire) => array_filter(
-                                            NotificationChannel::getAvailableEngagementOptions(),
-                                            function (string $channelValue) use ($livewire): bool {
-                                                $educatable = $livewire->getOwnerRecord();
-
-                                                assert($educatable instanceof Educatable);
-
-                                                $channel = NotificationChannel::tryFrom($channelValue);
-
-                                                if ($channel?->getCaseDisabled() ?? false) {
-                                                    return false;
-                                                }
-
-                                                return match ($channel) {
-                                                    NotificationChannel::Email => $educatable->hasValidEmail(),
-                                                    NotificationChannel::Sms => $educatable->hasValidSms(),
-                                                    default => true,
-                                                };
-                                            },
-                                            ARRAY_FILTER_USE_KEY
-                                        )
-                                    )
-                                    ->default(function (RelationManager $livewire): ?string {
-                                        $educatable = $livewire->getOwnerRecord();
-                                        assert($educatable instanceof Educatable);
-
-                                        return $educatable->getDefaultEngagementChannel()?->value;
-                                    })
-                                    ->live()
-                                    ->afterStateUpdated(function (mixed $state, RelationManager $livewire, Set $set) {
-                                        assert($livewire->getOwnerRecord() instanceof Educatable);
-
-                                        $channel = NotificationChannel::parse($state);
-                                        $educatable = $livewire->getOwnerRecord();
-
-                                        if (! $channel) {
-                                            $set('recipient_route_id', null);
-
-                                            return;
-                                        }
-
-                                        $route = $educatable->getDefaultRouteForEngagementChannel($channel);
-
-                                        $set('recipient_route_id', $route);
-                                    }),
-                                Select::make('recipient_route_id')
-                                    ->label(fn (Get $get): string => match (NotificationChannel::parse($get('channel'))) {
-                                        NotificationChannel::Email => 'Email address',
-                                        NotificationChannel::Sms => 'Phone number',
-                                    })
-                                    ->options(function (Get $get, RelationManager $livewire): array {
-                                        assert($livewire->getOwnerRecord() instanceof Student || $livewire->getOwnerRecord() instanceof Prospect);
-
-                                        return match (NotificationChannel::parse($get('channel'))) {
-                                            NotificationChannel::Email => $livewire->getOwnerRecord()->emailAddresses()
-                                                ->whereDoesntHave('bounced')
-                                                ->whereDoesntHave('optedOut', fn ($query) => $query->where('status', EmailAddressOptInOptOutStatus::OptedOut))
-                                                ->get()
-                                                ->mapWithKeys(fn (StudentEmailAddress | ProspectEmailAddress $emailAddress): array => [
-                                                    $emailAddress->getKey() => $emailAddress->address . (filled($emailAddress->type) ? " ({$emailAddress->type})" : ''),
-                                                ])
-                                                ->all(),
-                                            NotificationChannel::Sms => $livewire->getOwnerRecord()->phoneNumbers()
-                                                ->where('can_receive_sms', true)
-                                                ->whereDoesntHave('smsOptOut')
-                                                ->whereDoesntHave('bounced')
-                                                ->get()
-                                                ->mapWithKeys(fn (StudentPhoneNumber | ProspectPhoneNumber $phoneNumber): array => [
-                                                    $phoneNumber->getKey() => $phoneNumber->number . (filled($phoneNumber->ext) ? " (ext. {$phoneNumber->ext})" : '') . (filled($phoneNumber->type) ? " ({$phoneNumber->type})" : ''),
-                                                ])
-                                                ->all(),
-                                        };
-                                    })
-                                    ->default(function (RelationManager $livewire): ?string {
-                                        $educatable = $livewire->getOwnerRecord();
-
-                                        assert($educatable instanceof Educatable);
-
-                                        $channel = $educatable->getDefaultEngagementChannel();
-
-                                        if (! $channel) {
-                                            return null;
-                                        }
-
-                                        return $educatable->getDefaultRouteForEngagementChannel($channel);
-                                    })
-                                    ->required(),
-                            ]),
-                    ])
-                    ->columns(2),
-                Step::make('Message Details')
-                    ->schema([
-                        RichEditor::make('subject')
-                            ->label('Subject')
-                            ->toolbarButtons([])
-                            ->json()
-                            ->helperText('You may use "merge tags" to substitute information about a recipient into your subject line. Insert a "{{" in the subject line field to see a list of available merge tags')
-                            ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value)
-                            ->required()
-                            ->placeholder('Enter the email subject here...')
-                            ->columnSpanFull(),
-                        RichEditor::make('body')
-                            ->key('emailBody')
-                            ->fileAttachmentsDisk('s3-public')
-                            ->label('Body')
-                            ->toolbarButtons([
-                                ['bold', 'italic', 'link'],
-                                [ToolbarButtonGroup::make('Heading', ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'])->textualButtons(), 'bulletList', 'orderedList', 'horizontalRule'],
-                                ['textColor', 'small'],
-                                ['attachFiles', 'mergeTags'],
-                                ['clearFormatting'],
-                                ['undo', 'redo'],
-                            ])
-                            ->activePanel('mergeTags')
-                            ->resizableImages()
-                            ->json()
-                            ->required()
-                            ->hintAction(fn (RichEditor $component) => Action::make('loadEmailTemplate')
-                                ->schema([
-                                    Select::make('emailTemplate')
-                                        ->searchable()
-                                        ->options(function (Get $get): array {
-                                            return EmailTemplate::query()
-                                                ->when(
-                                                    $get('onlyMyTemplates'),
-                                                    fn (Builder $query) => $query->whereBelongsTo(auth()->user())
-                                                )
-                                                ->orderBy('name')
-                                                ->limit(50)
-                                                ->pluck('name', 'id')
-                                                ->toArray();
-                                        })
-                                        ->getSearchResultsUsing(function (Get $get, string $search): array {
-                                            return EmailTemplate::query()
-                                                ->when(
-                                                    $get('onlyMyTemplates'),
-                                                    fn (Builder $query) => $query->whereBelongsTo(auth()->user())
-                                                )
-                                                ->when(
-                                                    $get('onlyMyTeamTemplates'),
-                                                    fn (Builder $query) => $query->whereIn('user_id', auth()->user()->team->users->pluck('id'))
-                                                )
-                                                ->where(new Expression('lower(name)'), 'like', "%{$search}%")
-                                                ->orderBy('name')
-                                                ->limit(50)
-                                                ->pluck('name', 'id')
-                                                ->toArray();
-                                        })
-                                        ->getOptionLabelUsing(function (Get $get, string|int|null $value): ?string {
-                                            if (blank($value)) {
-                                                return null;
-                                            }
-
-                                            return EmailTemplate::query()
-                                                ->when(
-                                                    $get('onlyMyTemplates'),
-                                                    fn (Builder $query) => $query->whereBelongsTo(auth()->user())
-                                                )
-                                                ->when(
-                                                    $get('onlyMyTeamTemplates'),
-                                                    fn (Builder $query) => $query->whereIn('user_id', auth()->user()->team->users->pluck('id'))
-                                                )
-                                                ->whereKey($value)
-                                                ->value('name');
-                                        }),
-                                    Checkbox::make('onlyMyTemplates')
-                                        ->label('Only show my templates')
-                                        ->live()
-                                        ->afterStateUpdated(fn (Set $set) => $set('emailTemplate', null)),
-                                    Checkbox::make('onlyMyTeamTemplates')
-                                        ->label("Only show my team's templates")
-                                        ->live()
-                                        ->afterStateUpdated(fn (Set $set) => $set('emailTemplate', null)),
-                                ])
-                                ->action(function (array $data) use ($component) {
-                                    $template = EmailTemplate::find($data['emailTemplate']);
-
-                                    if (! $template) {
-                                        return;
-                                    }
-
-                                    $component->state($template->content);
-                                }))
-                            ->getFileAttachmentUrlFromAnotherRecordUsing(function (mixed $file): ?string {
-                                return Media::query()
-                                    ->where('uuid', $file)
-                                    ->where('model_type', (new EmailTemplate())->getMorphClass())
-                                    ->first()
-                                    ?->getUrl();
-                            })
-                            ->saveFileAttachmentFromAnotherRecordUsing(function (mixed $file, ?Engagement $record): ?string {
-                                if (! $record) {
-                                    return null;
-                                }
-
-                                return Media::query()
-                                    ->where('uuid', $file)
-                                    ->where('model_type', (new EmailTemplate())->getMorphClass())
-                                    ->first()
-                                    ?->copy($record, 'body', 's3-public')
-                                    ->uuid;
-                            })
-                            ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value)
-                            ->helperText('You can insert recipient or your information by typing {{ and choosing a merge value to insert.')
-                            ->columnSpanFull(),
-                        EngagementSmsBodyInput::make(context: 'create'),
-                        Actions::make([
-                            RelationManagerDraftWithAiAction::make()
-                                ->mergeTags(Engagement::getMergeTags()),
-                        ]),
-                    ]),
-                Step::make('Email Signature')
-                    ->schema([
-                        Toggle::make('is_signature_enabled')
-                            ->label('Include signature')
-                            ->helperText('You may configure your email signature in Profile Settings by selecting your avatar in the upper right portion of the screen.')
-                            ->live(),
-                        RichEditor::make('signature')
-                            ->toolbarButtons([['bold', 'italic', 'strike', 'underline', 'small', 'textColor', 'link', 'attachFiles']])
-                            ->extraInputAttributes(['style' => 'min-height: 12rem;'])
-                            ->json()
-                            ->required(fn (Get $get) => $get('is_signature_enabled'))
-                            ->fileAttachmentsDisk('s3-public')
-                            ->visible(fn (Get $get) => $get('is_signature_enabled'))
-                            ->default(auth()->user()->signature)
-                            ->saveRelationshipsUsing(null),
-                    ])
-                    ->visible(auth()->user()->is_signature_enabled)
-                    ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value),
-                Step::make('Delivery Details')
-                    ->schema([
-                        Toggle::make('send_later')
-                            ->reactive()
-                            ->helperText('By default, this message will send as soon as it is created unless you schedule it to send later.'),
-                        DateTimePicker::make('scheduled_at')
-                            ->required()
-                            ->visible(fn (Get $get) => $get('send_later')),
-                    ]),
-            ])
-            ->action(function (array $data, Schema $schema, RelationManager $livewire) {
-                $recipient = $livewire->getOwnerRecord();
-
-                throw_if(
-                    ! $recipient instanceof Educatable,
-                    new Exception('Recipient must implement Educatable interface.'),
-                );
-
-                $data['subject'] ??= ['type' => 'doc', 'content' => []];
-                $data['subject']['content'] = [
-                    ...($data['subject']['content'] ?? []),
-                ];
-
-                $data['body'] ??= ['type' => 'doc', 'content' => []];
-                $data['body']['content'] = [
-                    ...($data['body']['content'] ?? []),
-                    ...($data['signature']['content'] ?? []),
-                ];
-
-                $channel = NotificationChannel::parse($data['channel']);
-
-                $recipientRoute = match ($channel) {
-                    NotificationChannel::Email => $recipient->emailAddresses()->find($data['recipient_route_id'] ?? null)?->address,
-                    NotificationChannel::Sms => $recipient->phoneNumbers()->find($data['recipient_route_id'] ?? null)?->number,
-                    default => null,
-                };
-
-                $engagement = app(CreateEngagement::class)->execute(new EngagementCreationData(
-                    user: auth()->user(),
-                    recipient: $recipient,
-                    channel: $channel,
-                    subject: $data['subject'] ?? null,
-                    body: $data['body'] ?? null,
-                    scheduledAt: ($data['send_later'] ?? false) ? Carbon::parse($data['scheduled_at'] ?? null) : null,
-                    recipientRoute: $recipientRoute,
-                    schema: $schema,
-                ));
-            })
-            ->modalSubmitActionLabel('Send')
-            ->modalCloseButton(false)
-            ->closeModalByClickingAway(false)
-            ->closeModalByEscaping(false)
-            ->createAnother(false);
-    }
-
-    public static function getDefaultName(): ?string
-    {
-        return 'engage';
-    }
-}
