+++ app-modules/engagement/src/Filament/Actions/SendEngagementAction.php
@@ -39,28 +39,30 @@ namespace AdvisingApp\Engagement\Filament\Actions;
 use AdvisingApp\Engagement\Actions\CreateEngagement;
 use AdvisingApp\Engagement\DataTransferObjects\EngagementCreationData;
 use AdvisingApp\Engagement\Filament\Forms\Components\EngagementSmsBodyInput;
-use AdvisingApp\Engagement\Models\EmailTemplate;
+use AdvisingApp\Engagement\Filament\Schemas\Components\EngagementEmailBodyInput;
+use AdvisingApp\Engagement\Filament\Schemas\Components\EngagementScheduledAtDateTimePicker;
+use AdvisingApp\Engagement\Filament\Schemas\Components\EngagementSendLaterToggle;
+use AdvisingApp\Engagement\Filament\Schemas\Components\EngagementSubjectInput;
 use AdvisingApp\Engagement\Models\Engagement;
 use AdvisingApp\Notification\Enums\NotificationChannel;
 use AdvisingApp\Prospect\Models\Prospect;
 use AdvisingApp\Prospect\Models\ProspectEmailAddress;
 use AdvisingApp\Prospect\Models\ProspectPhoneNumber;
 use AdvisingApp\StudentDataModel\Enums\EmailAddressOptInOptOutStatus;
+use AdvisingApp\StudentDataModel\Models\Contracts\Educatable;
 use AdvisingApp\StudentDataModel\Models\Student;
 use AdvisingApp\StudentDataModel\Models\StudentEmailAddress;
 use AdvisingApp\StudentDataModel\Models\StudentPhoneNumber;
 use App\Filament\Forms\Components\EducatableSelect;
+use Closure;
 use Exception;
 use Filament\Actions\Action;
-use Filament\Forms\Components\Checkbox;
-use Filament\Forms\Components\DateTimePicker;
 use Filament\Forms\Components\RichEditor;
-use Filament\Forms\Components\RichEditor\ToolbarButtonGroup;
 use Filament\Forms\Components\Select;
 use Filament\Forms\Components\Toggle;
 use Filament\Forms\Components\ToggleButtons;
-use Filament\Pages\Page;
 use Filament\Schemas\Components\Actions;
+use Filament\Schemas\Components\Component;
 use Filament\Schemas\Components\Fieldset;
 use Filament\Schemas\Components\Grid;
 use Filament\Schemas\Components\Text;
@@ -68,21 +70,18 @@ use Filament\Schemas\Components\Utilities\Get;
 use Filament\Schemas\Components\Utilities\Set;
 use Filament\Schemas\Components\Wizard\Step;
 use Filament\Schemas\Schema;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Query\Expression;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Str;
-use Spatie\MediaLibrary\MediaCollections\Models\Media;
 
 class SendEngagementAction extends Action
 {
-    protected Student | Prospect | null $educatable = null;
+    protected Student | Prospect | Closure | null $educatable = null;
+
+    protected ?Closure $draftWithAiActionUsing = null;
 
     protected function setUp(): void
     {
         parent::setUp();
-
-        $action = $this;
 
         $this->icon('heroicon-m-chat-bubble-bottom-center-text')
             ->slideOver()
@@ -128,8 +127,12 @@ class SendEngagementAction extends Action
 
                 return null;
             })
-            ->mountUsing(function (array $arguments, Schema $schema, Page $livewire) {
-                $livewire->dispatch('engage-action-finished-loading');
+            ->mountUsing(function (array $arguments, Schema $schema) {
+                $livewire = $this->getLivewire();
+
+                if (method_exists($livewire, 'dispatch')) {
+                    $livewire->dispatch('engage-action-finished-loading');
+                }
 
                 $educatable = $this->getEducatable();
 
@@ -153,348 +156,9 @@ class SendEngagementAction extends Action
                     ]);
                 }
             })
-            ->steps(fn (): array => [
-                Step::make('Recipient Details')
-                    ->schema([
-                        ...$this->getEducatable() ? [] : [
-                            EducatableSelect::make(
-                                name: 'recipient',
-                                isExcludingConvertedProspects: true,
-                                modifyKeySelectUsing: function (Select $select): Select {
-                                    return $select->disableOptionWhen(function (string $value): bool {
-                                        static $noContactCache = [];
-                                        $cacheKey = $value;
-
-                                        if (! array_key_exists($cacheKey, $noContactCache)) {
-                                            $educatable = Student::find($value) ?? Prospect::find($value);
-                                            $noContactCache[$cacheKey] = $educatable
-                                                ? ! $educatable->hasAnyValidContactRoute()
-                                                : false;
-                                        }
-
-                                        return $noContactCache[$cacheKey];
-                                    });
-                                },
-                            )
-                                ->label('Recipient Info')
-                                ->live()
-                                ->typeSelectToggleButtons()
-                                ->required()
-                                ->columns(2)
-                                ->afterStateUpdated(function (Get $get, Set $set) {
-                                    $educatable = match ($get('recipient_type')) {
-                                        'student' => Student::find($get('recipient_id')),
-                                        'prospect' => Str::isUuid($get('recipient_id')) ? Prospect::find($get('recipient_id')) : null,
-                                        default => null,
-                                    };
-
-                                    if (! $educatable) {
-                                        return;
-                                    }
-
-                                    $defaultChannel = $educatable->getDefaultEngagementChannel();
-
-                                    if ($defaultChannel) {
-                                        $set('channel', $defaultChannel->value);
-                                        $set('recipient_route_id', $educatable->getDefaultRouteForEngagementChannel($defaultChannel));
-
-                                        return;
-                                    }
-                                }),
-                        ],
-                        Grid::make(2)
-                            ->schema(function (Get $get) use ($action): array {
-                                return [
-                                    Fieldset::make('Message Type')
-                                        ->schema([
-                                            ToggleButtons::make('channel')
-                                                ->options(function (Get $get) use ($action): array {
-                                                    $educatable = $action->resolveEducatable($get);
-
-                                                    if (! $educatable) {
-                                                        return [];
-                                                    }
-
-                                                    return array_filter(
-                                                        NotificationChannel::getAvailableEngagementOptions(),
-                                                        function (string $channelValue) use ($educatable): bool {
-                                                            $channel = NotificationChannel::tryFrom($channelValue);
-
-                                                            if ($channel?->getCaseDisabled() ?? false) {
-                                                                return false;
-                                                            }
-
-                                                            return match ($channel) {
-                                                                NotificationChannel::Email => $educatable->hasValidEmail(),
-                                                                NotificationChannel::Sms => $educatable->hasValidSms(),
-                                                                default => true,
-                                                            };
-                                                        },
-                                                        ARRAY_FILTER_USE_KEY
-                                                    );
-                                                })
-                                                ->default(NotificationChannel::Email->value)
-                                                ->inline()
-                                                ->live()
-                                                ->afterStateUpdated(function (mixed $state, Get $get, Set $set) use ($action): void {
-                                                    $educatable = $action->resolveEducatable($get);
-
-                                                    if (! $educatable) {
-                                                        return;
-                                                    }
-
-                                                    $channel = NotificationChannel::parse($state);
-                                                    $set('recipient_route_id', $educatable->getDefaultRouteForEngagementChannel($channel));
-                                                }),
-                                            Select::make('recipient_route_id')
-                                                ->label(fn (Get $get): string => match (NotificationChannel::parse($get('channel'))) {
-                                                    NotificationChannel::Email => 'Email address',
-                                                    NotificationChannel::Sms => 'Phone number',
-                                                    default => throw new Exception('Invalid channel.'),
-                                                })
-                                                ->options(function (Get $get) use ($action): array {
-                                                    $educatable = $action->resolveEducatable($get);
-
-                                                    if (! $educatable) {
-                                                        return [];
-                                                    }
-
-                                                    return match (NotificationChannel::parse($get('channel'))) {
-                                                        NotificationChannel::Email => $educatable->emailAddresses()
-                                                            ->whereDoesntHave('bounced')
-                                                            ->whereDoesntHave('optedOut', fn ($query) => $query->where('status', EmailAddressOptInOptOutStatus::OptedOut))
-                                                            ->get()
-                                                            ->mapWithKeys(fn (StudentEmailAddress | ProspectEmailAddress $emailAddress): array => [
-                                                                $emailAddress->getKey() => $emailAddress->address . (filled($emailAddress->type) ? " ({$emailAddress->type})" : ''),
-                                                            ])
-                                                            ->all(),
-                                                        NotificationChannel::Sms => $educatable->phoneNumbers()
-                                                            ->where('can_receive_sms', true)
-                                                            ->whereDoesntHave('smsOptOut')
-                                                            ->whereDoesntHave('bounced')
-                                                            ->get()
-                                                            ->mapWithKeys(fn (StudentPhoneNumber | ProspectPhoneNumber $phoneNumber): array => [
-                                                                $phoneNumber->getKey() => $phoneNumber->number . (filled($phoneNumber->ext) ? " (ext. {$phoneNumber->ext})" : '') . (filled($phoneNumber->type) ? " ({$phoneNumber->type})" : ''),
-                                                            ])
-                                                            ->all(),
-                                                        default => [],
-                                                    };
-                                                })
-                                                ->disabled(function (Get $get) use ($action): bool {
-                                                    $educatable = $action->resolveEducatable($get);
-
-                                                    return blank($educatable);
-                                                })
-                                                ->required(),
-                                        ])
-                                        ->visible(function (Get $get) use ($action): bool {
-                                            $educatable = $action->resolveEducatable($get);
-
-                                            return $educatable !== null && $educatable->hasAnyValidContactRoute();
-                                        })
-                                        ->columnSpanFull(),
-                                    Text::make('This recipient does not have valid contact information. Please select a different recipient.')
-                                        ->visible(function (Get $get) use ($action): bool {
-                                            $educatable = $action->resolveEducatable($get);
-
-                                            return $educatable !== null && ! $educatable->hasAnyValidContactRoute();
-                                        })
-                                        ->columnSpanFull(),
-                                ];
-                            })
-                            ->visible(fn (Get $get) => ! is_null($action->getEducatable()) || $get('recipient_id')),
-                    ]),
-                Step::make('Message Details')
-                    ->schema(function (Get $get): array {
-                        $educatable = $this->resolveEducatable($get);
-
-                        return [
-                            RichEditor::make('subject')
-                                ->label('Subject')
-                                ->toolbarButtons([])
-                                ->json()
-                                ->helperText('You may use "merge tags" to substitute information about a student into your subject line. Insert a "{{" in the subject line field to see a list of available merge tags')
-                                ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value)
-                                ->required()
-                                ->placeholder('Enter the email subject here...')
-                                ->columnSpanFull(),
-                            RichEditor::make('body')
-                                ->key('emailBody')
-                                ->fileAttachmentsDisk('s3-public')
-                                ->label('Body')
-                                ->toolbarButtons([
-                                    ['bold', 'italic', 'link'],
-                                    [ToolbarButtonGroup::make('Heading', ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'])->textualButtons(), 'bulletList', 'orderedList', 'horizontalRule'],
-                                    ['textColor', 'small'],
-                                    ['attachFiles', 'mergeTags'],
-                                    ['clearFormatting'],
-                                    ['undo', 'redo'],
-                                ])
-                                ->activePanel('mergeTags')
-                                ->resizableImages()
-                                ->json()
-                                ->required()
-                                ->hintAction(fn (RichEditor $component) => Action::make('loadEmailTemplate')
-                                    ->schema([
-                                        Select::make('emailTemplate')
-                                            ->searchable()
-                                            ->options(function (Get $get): array {
-                                                return EmailTemplate::query()
-                                                    ->when(
-                                                        $get('onlyMyTemplates'),
-                                                        fn (Builder $query) => $query->whereBelongsTo(auth()->user())
-                                                    )
-                                                    ->orderBy('name')
-                                                    ->limit(50)
-                                                    ->pluck('name', 'id')
-                                                    ->toArray();
-                                            })
-                                            ->getSearchResultsUsing(function (Get $get, string $search): array {
-                                                return EmailTemplate::query()
-                                                    ->when(
-                                                        $get('onlyMyTemplates'),
-                                                        fn (Builder $query) => $query->whereBelongsTo(auth()->user())
-                                                    )
-                                                    ->when(
-                                                        $get('onlyMyTeamTemplates'),
-                                                        fn (Builder $query) => $query->whereIn('user_id', auth()->user()->team->users()->pluck('id'))
-                                                    )
-                                                    ->where(new Expression('lower(name)'), 'like', "%{$search}%")
-                                                    ->orderBy('name')
-                                                    ->limit(50)
-                                                    ->pluck('name', 'id')
-                                                    ->toArray();
-                                            })
-                                            ->getOptionLabelUsing(function (Get $get, string|int|null $value): ?string {
-                                                if (blank($value)) {
-                                                    return null;
-                                                }
-
-                                                return EmailTemplate::query()
-                                                    ->when(
-                                                        $get('onlyMyTemplates'),
-                                                        fn (Builder $query) => $query->whereBelongsTo(auth()->user())
-                                                    )
-                                                    ->when(
-                                                        $get('onlyMyTeamTemplates'),
-                                                        fn (Builder $query) => $query->whereIn('user_id', auth()->user()->team->users()->pluck('id'))
-                                                    )
-                                                    ->whereKey($value)
-                                                    ->value('name');
-                                            }),
-                                        Checkbox::make('onlyMyTemplates')
-                                            ->label('Only show my templates')
-                                            ->live()
-                                            ->afterStateUpdated(fn (Set $set) => $set('emailTemplate', null)),
-                                        Checkbox::make('onlyMyTeamTemplates')
-                                            ->label("Only show my team's templates")
-                                            ->live()
-                                            ->afterStateUpdated(fn (Set $set) => $set('emailTemplate', null)),
-                                    ])
-                                    ->action(function (array $data) use ($component) {
-                                        $template = EmailTemplate::find($data['emailTemplate']);
-
-                                        if (! $template) {
-                                            return;
-                                        }
-
-                                        $component->state($template->content);
-                                    }))
-                                ->getFileAttachmentUrlFromAnotherRecordUsing(function (mixed $file): ?string {
-                                    return Media::query()
-                                        ->where('uuid', $file)
-                                        ->where('model_type', (new EmailTemplate())->getMorphClass())
-                                        ->first()
-                                        ?->getUrl();
-                                })
-                                ->saveFileAttachmentFromAnotherRecordUsing(function (mixed $file, ?Engagement $record): ?string {
-                                    if (! $record) {
-                                        return null;
-                                    }
-
-                                    return Media::query()
-                                        ->where('uuid', $file)
-                                        ->where('model_type', (new EmailTemplate())->getMorphClass())
-                                        ->first()
-                                        ?->copy($record, 'body', 's3-public')
-                                        ->uuid;
-                                })
-                                ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value)
-                                ->helperText('You can insert student or your information by typing {{ and choosing a merge value to insert.')
-                                ->columnSpanFull(),
-                            EngagementSmsBodyInput::make(context: 'create'),
-                            ...$educatable ? [Actions::make([
-                                DraftWithAiAction::make()
-                                    ->mergeTags(Engagement::getMergeTags())
-                                    ->educatable($educatable),
-                            ])] : [],
-                        ];
-                    }),
-                Step::make('Email Signature')
-                    ->schema([
-                        Toggle::make('is_signature_enabled')
-                            ->label('Include Signature')
-                            ->helperText('You may configure your email signature in Profile Settings by selecting your avatar in the upper right portion of the screen.')
-                            ->live(),
-                        RichEditor::make('signature')
-                            ->toolbarButtons([['bold', 'italic', 'strike', 'underline', 'small', 'textColor', 'link', 'attachFiles']])
-                            ->extraInputAttributes(['style' => 'min-height: 12rem;'])
-                            ->json()
-                            ->required(fn (Get $get) => $get('is_signature_enabled'))
-                            ->fileAttachmentsDisk('s3-public')
-                            ->visible(fn (Get $get) => $get('is_signature_enabled'))
-                            ->default(auth()->user()->signature)
-                            ->saveRelationshipsUsing(null),
-                    ])
-                    ->visible(auth()->user()->is_signature_enabled)
-                    ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value),
-                Step::make('Delivery Details')
-                    ->schema([
-                        Toggle::make('send_later')
-                            ->reactive()
-                            ->helperText('By default, this message will send as soon as it is created unless you schedule it to send later.'),
-                        DateTimePicker::make('scheduled_at')
-                            ->required()
-                            ->visible(fn (Get $get) => $get('send_later')),
-                    ]),
-            ])
-            ->action(function (array $data, Schema $schema, Page $livewire) {
-                /** @var Student | Prospect $recipient */
-                $recipient = $this->getEducatable() ?? match ($data['recipient_type']) {
-                    'student' => Student::find($data['recipient_id']),
-                    'prospect' => Str::isUuid($data['recipient_id']) ? Prospect::find($data['recipient_id']) : null,
-                    default => null,
-                };
-                $data['subject'] ??= ['type' => 'doc', 'content' => []];
-                $data['subject']['content'] = [
-                    ...($data['subject']['content'] ?? []),
-                ];
-                $data['body'] ??= ['type' => 'doc', 'content' => []];
-                $data['body']['content'] = [
-                    ...($data['body']['content'] ?? []),
-                    ...($data['signature']['content'] ?? []),
-                ];
-
-                $channel = NotificationChannel::parse($data['channel']);
-
-                $recipientRoute = match ($channel) {
-                    NotificationChannel::Email => $recipient->emailAddresses()->find($data['recipient_route_id'] ?? null)?->address,
-                    NotificationChannel::Sms => $recipient->phoneNumbers()->find($data['recipient_route_id'] ?? null)?->number,
-                    default => null,
-                };
-
-                $engagement = app(CreateEngagement::class)->execute(new EngagementCreationData(
-                    user: auth()->user(),
-                    recipient: $recipient,
-                    channel: $channel,
-                    subject: $data['subject'] ?? null,
-                    body: $data['body'] ?? null,
-                    scheduledAt: ($data['send_later'] ?? false) ? Carbon::parse($data['scheduled_at'] ?? null) : null,
-                    recipientRoute: $recipientRoute,
-                    schema: $schema,
-                ));
-
-                $livewire->dispatch('engagement-sent');
+            ->steps(fn (): array => $this->getEngagementSteps())
+            ->action(function (array $data, Schema $schema) {
+                $this->createEngagement($data, $schema);
             })
             ->modalSubmitActionLabel('Send')
             ->modalCloseButton(false)
@@ -507,7 +171,7 @@ class SendEngagementAction extends Action
         return 'engage';
     }
 
-    public function educatable(Student | Prospect | null $educatable): static
+    public function educatable(Student | Prospect | Closure | null $educatable): static
     {
         $this->educatable = $educatable;
 
@@ -516,6 +180,10 @@ class SendEngagementAction extends Action
 
     public function getEducatable(): Student | Prospect | null
     {
+        if ($this->educatable instanceof Closure) {
+            return ($this->educatable)();
+        }
+
         return $this->educatable;
     }
 
@@ -526,5 +194,271 @@ class SendEngagementAction extends Action
             'prospect' => Str::isUuid($get('recipient_id')) ? Prospect::find($get('recipient_id')) : null,
             default => null,
         };
+    }
+
+    public function draftWithAiAction(Closure $callback): static
+    {
+        $this->draftWithAiActionUsing = $callback;
+
+        return $this;
+    }
+
+    /**
+     * @return array<string>
+     */
+    public static function getDefaultMergeTags(): array
+    {
+        return Engagement::getMergeTags();
+    }
+
+    /**
+     * @return array<Component>
+     */
+    protected function getEngagementSteps(): array
+    {
+        return [
+            Step::make('Recipient Details')
+                ->schema([
+                    ...$this->getEducatable() ? [] : [
+                        EducatableSelect::make(
+                            name: 'recipient',
+                            isExcludingConvertedProspects: true,
+                            modifyKeySelectUsing: function (Select $select): Select {
+                                return $select->disableOptionWhen(function (string $value): bool {
+                                    static $noContactCache = [];
+                                    $cacheKey = $value;
+
+                                    if (! array_key_exists($cacheKey, $noContactCache)) {
+                                        $educatable = Student::find($value) ?? Prospect::find($value);
+                                        $noContactCache[$cacheKey] = $educatable
+                                            ? ! $educatable->hasAnyValidContactRoute()
+                                            : false;
+                                    }
+
+                                    return $noContactCache[$cacheKey];
+                                });
+                            },
+                        )
+                            ->label('Recipient Info')
+                            ->live()
+                            ->typeSelectToggleButtons()
+                            ->required()
+                            ->columns(2)
+                            ->afterStateUpdated(function (Get $get, Set $set) {
+                                $educatable = match ($get('recipient_type')) {
+                                    'student' => Student::find($get('recipient_id')),
+                                    'prospect' => Str::isUuid($get('recipient_id')) ? Prospect::find($get('recipient_id')) : null,
+                                    default => null,
+                                };
+
+                                if (! $educatable) {
+                                    return;
+                                }
+
+                                $defaultChannel = $educatable->getDefaultEngagementChannel();
+
+                                if ($defaultChannel) {
+                                    $set('channel', $defaultChannel->value);
+                                    $set('recipient_route_id', $educatable->getDefaultRouteForEngagementChannel($defaultChannel));
+
+                                    return;
+                                }
+                            }),
+                    ],
+                    Grid::make(2)
+                        ->schema(function (Get $get): array {
+                            return [
+                                Fieldset::make('Message Type')
+                                    ->schema([
+                                        ToggleButtons::make('channel')
+                                            ->options(function (Get $get): array {
+                                                $educatable = $this->resolveEducatable($get);
+
+                                                if (! $educatable) {
+                                                    return [];
+                                                }
+
+                                                return array_filter(
+                                                    NotificationChannel::getAvailableEngagementOptions(),
+                                                    function (string $channelValue) use ($educatable): bool {
+                                                        $channel = NotificationChannel::tryFrom($channelValue);
+
+                                                        if ($channel?->getCaseDisabled() ?? false) {
+                                                            return false;
+                                                        }
+
+                                                        return match ($channel) {
+                                                            NotificationChannel::Email => $educatable->hasValidEmail(),
+                                                            NotificationChannel::Sms => $educatable->hasValidSms(),
+                                                            default => true,
+                                                        };
+                                                    },
+                                                    ARRAY_FILTER_USE_KEY
+                                                );
+                                            })
+                                            ->default(NotificationChannel::Email->value)
+                                            ->inline()
+                                            ->live()
+                                            ->afterStateUpdated(function (mixed $state, Get $get, Set $set): void {
+                                                $educatable = $this->resolveEducatable($get);
+
+                                                if (! $educatable) {
+                                                    return;
+                                                }
+
+                                                $channel = NotificationChannel::parse($state);
+                                                $set('recipient_route_id', $educatable->getDefaultRouteForEngagementChannel($channel));
+                                            }),
+                                        Select::make('recipient_route_id')
+                                            ->label(fn (Get $get): string => match (NotificationChannel::parse($get('channel'))) {
+                                                NotificationChannel::Email => 'Email address',
+                                                NotificationChannel::Sms => 'Phone number',
+                                                default => throw new Exception('Invalid channel.'),
+                                            })
+                                            ->options(function (Get $get): array {
+                                                $educatable = $this->resolveEducatable($get);
+
+                                                if (! $educatable) {
+                                                    return [];
+                                                }
+
+                                                return match (NotificationChannel::parse($get('channel'))) {
+                                                    NotificationChannel::Email => $educatable->emailAddresses()
+                                                        ->whereDoesntHave('bounced')
+                                                        ->whereDoesntHave('optedOut', fn ($query) => $query->where('status', EmailAddressOptInOptOutStatus::OptedOut))
+                                                        ->get()
+                                                        ->mapWithKeys(fn (StudentEmailAddress | ProspectEmailAddress $emailAddress): array => [
+                                                            $emailAddress->getKey() => $emailAddress->address . (filled($emailAddress->type) ? " ({$emailAddress->type})" : ''),
+                                                        ])
+                                                        ->all(),
+                                                    NotificationChannel::Sms => $educatable->phoneNumbers()
+                                                        ->where('can_receive_sms', true)
+                                                        ->whereDoesntHave('smsOptOut')
+                                                        ->whereDoesntHave('bounced')
+                                                        ->get()
+                                                        ->mapWithKeys(fn (StudentPhoneNumber | ProspectPhoneNumber $phoneNumber): array => [
+                                                            $phoneNumber->getKey() => $phoneNumber->number . (filled($phoneNumber->ext) ? " (ext. {$phoneNumber->ext})" : '') . (filled($phoneNumber->type) ? " ({$phoneNumber->type})" : ''),
+                                                        ])
+                                                        ->all(),
+                                                    default => [],
+                                                };
+                                            })
+                                            ->disabled(function (Get $get): bool {
+                                                $educatable = $this->resolveEducatable($get);
+
+                                                return blank($educatable);
+                                            })
+                                            ->required(),
+                                    ])
+                                    ->visible(function (Get $get): bool {
+                                        $educatable = $this->resolveEducatable($get);
+
+                                        return $educatable !== null && $educatable->hasAnyValidContactRoute();
+                                    })
+                                    ->columnSpanFull(),
+                                Text::make('This recipient does not have valid contact information. Please select a different recipient.')
+                                    ->visible(function (Get $get): bool {
+                                        $educatable = $this->resolveEducatable($get);
+
+                                        return $educatable !== null && ! $educatable->hasAnyValidContactRoute();
+                                    })
+                                    ->columnSpanFull(),
+                            ];
+                        })
+                        ->visible(fn (Get $get) => ! is_null($this->getEducatable()) || $get('recipient_id')),
+                ]),
+            Step::make('Message Details')
+                ->schema(function (Get $get): array {
+                    $educatable = $this->resolveEducatable($get);
+
+                    return [
+                        EngagementSubjectInput::make()
+                            ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value),
+                        EngagementEmailBodyInput::make()
+                            ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value),
+                        EngagementSmsBodyInput::make(context: 'create'),
+                        ...$educatable ? [Actions::make([
+                            $this->getDraftWithAiAction($educatable),
+                        ])] : [],
+                    ];
+                }),
+            Step::make('Email Signature')
+                ->schema([
+                    Toggle::make('is_signature_enabled')
+                        ->label('Include Signature')
+                        ->helperText('You may configure your email signature in Profile Settings by selecting your avatar in the upper right portion of the screen.')
+                        ->live(),
+                    RichEditor::make('signature')
+                        ->toolbarButtons([['bold', 'italic', 'strike', 'underline', 'small', 'textColor', 'link', 'attachFiles']])
+                        ->extraInputAttributes(['style' => 'min-height: 12rem;'])
+                        ->json()
+                        ->required(fn (Get $get) => $get('is_signature_enabled'))
+                        ->fileAttachmentsDisk('s3-public')
+                        ->visible(fn (Get $get) => $get('is_signature_enabled'))
+                        ->default(auth()->user()->signature)
+                        ->saveRelationshipsUsing(null),
+                ])
+                ->visible(auth()->user()->is_signature_enabled)
+                ->hidden(fn (Get $get): bool => $get('channel') === NotificationChannel::Sms->value),
+            Step::make('Delivery Details')
+                ->schema([
+                    EngagementSendLaterToggle::make(),
+                    EngagementScheduledAtDateTimePicker::make(),
+                ]),
+        ];
+    }
+
+    /**
+     * @param  array<string, mixed>  $data
+     */
+    protected function createEngagement(array $data, Schema $schema): void
+    {
+        /** @var Student | Prospect $recipient */
+        $recipient = $this->getEducatable() ?? match ($data['recipient_type']) {
+            'student' => Student::find($data['recipient_id']),
+            'prospect' => Str::isUuid($data['recipient_id']) ? Prospect::find($data['recipient_id']) : null,
+            default => null,
+        };
+        $data['subject'] ??= ['type' => 'doc', 'content' => []];
+        $data['subject']['content'] = [
+            ...($data['subject']['content'] ?? []),
+        ];
+        $data['body'] ??= ['type' => 'doc', 'content' => []];
+        $data['body']['content'] = [
+            ...($data['body']['content'] ?? []),
+            ...($data['signature']['content'] ?? []),
+        ];
+
+        $channel = NotificationChannel::parse($data['channel']);
+
+        $recipientRoute = match ($channel) {
+            NotificationChannel::Email => $recipient->emailAddresses()->find($data['recipient_route_id'] ?? null)?->address,
+            NotificationChannel::Sms => $recipient->phoneNumbers()->find($data['recipient_route_id'] ?? null)?->number,
+            default => null,
+        };
+
+        app(CreateEngagement::class)->execute(new EngagementCreationData(
+            user: auth()->user(),
+            recipient: $recipient,
+            channel: $channel,
+            subject: $data['subject'] ?? null,
+            body: $data['body'] ?? null,
+            scheduledAt: ($data['send_later'] ?? false) ? Carbon::parse($data['scheduled_at'] ?? null) : null,
+            recipientRoute: $recipientRoute,
+            schema: $schema,
+        ));
+
+        $this->getLivewire()->dispatch('engagement-sent');
+    }
+
+    protected function getDraftWithAiAction(Educatable $educatable): Action
+    {
+        if ($this->draftWithAiActionUsing) {
+            return $this->evaluate($this->draftWithAiActionUsing);
+        }
+
+        return DraftWithAiAction::make()
+            ->educatable($educatable)
+            ->mergeTags(static::getDefaultMergeTags());
     }
 }
